# docs/openapi.yaml
openapi: 3.0.0
info:
  title: GoSpeedTest API
  description: High-performance page speed analysis engine written in Go.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Job:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, PARTIAL, FAILED]
        tiers:
          type: array
          items:
            type: string
        runs:
          type: integer
        webhook_url:
          type: string
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        budget:
          $ref: '#/components/schemas/Budget'
        budget_result:
          $ref: '#/components/schemas/BudgetEvaluation'
        schedule_id:
          type: string
          description: Set when the job was created by a schedule.
        profile:
          $ref: '#/components/schemas/ThrottlingProfile'

    Budget:
      type: object
      description: >
        Performance budget: assertions keyed by metric name
        (e.g. network.ttfb_ms, network.total_ms, network.dns_lookup_ms,
        network.tls_handshake_ms, network.response_bytes, browser.page_load_ms,
        browser.dom_content_loaded_ms, browser.resource_count, vitals.lcp_ms,
        vitals.fcp_ms, lighthouse.performance, lighthouse.accessibility,
        lighthouse.best_practices, lighthouse.seo).
      required:
        - assertions
      properties:
        assertions:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/BudgetAssertion'
      example:
        assertions:
          network.ttfb_ms: { max: 500 }
          vitals.lcp_ms: { max: 2500, level: warn }
          lighthouse.performance: { min: 0.9 }

    BudgetAssertion:
      type: object
      description: Threshold on one metric. At least one of max/min is required.
      properties:
        max:
          type: number
        min:
          type: number
        level:
          type: string
          enum: [warn, error]
          default: error

    BudgetEvaluation:
      type: object
      description: >
        Outcome of evaluating the budget against the median of the job's runs.
        Evaluated at job completion; never affects the job status itself.
      properties:
        status:
          type: string
          enum: [pass, warn, fail]
        assertions:
          type: array
          items:
            type: object
            properties:
              metric: { type: string }
              level: { type: string, enum: [warn, error] }
              op: { type: string, enum: [max, min] }
              threshold: { type: number }
              actual:
                type: number
                nullable: true
                description: null when the metric was never collected (that trips the assertion).
              status: { type: string, enum: [pass, warn, fail] }

    MetricStats:
      type: object
      properties:
        count: { type: integer }
        avg: { type: number }
        p50: { type: number }
        p75: { type: number }
        p95: { type: number }

    History:
      type: object
      properties:
        url: { type: string }
        test_count: { type: integer }
        avg_ttfb_ms:
          type: number
          description: Legacy field, equals metrics["network.ttfb_ms"].avg.
        avg_total_ms:
          type: number
          description: Legacy field, equals metrics["network.total_ms"].avg.
        metrics:
          type: object
          description: Keyed by budget metric names (e.g. network.ttfb_ms, vitals.lcp_ms).
          additionalProperties:
            $ref: '#/components/schemas/MetricStats'

    Schedule:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
        tiers:
          type: array
          items: { type: string }
        runs: { type: integer }
        interval_seconds: { type: integer, minimum: 60 }
        budget:
          $ref: '#/components/schemas/Budget'
        webhook_url: { type: string }
        profile:
          $ref: '#/components/schemas/ThrottlingProfile'
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        last_run_at: { type: string, format: date-time }
        next_run_at: { type: string, format: date-time }

    ThrottlingProfile:
      type: string
      enum: [none, 4g, fast-3g, slow-3g]
      default: none
      description: >
        Network/CPU throttling applied to browser-driven tiers (browser,
        vitals), mirroring Chrome DevTools presets. The network tier uses a
        plain Go HTTP client and is never throttled.

    CreateScheduleRequest:
      type: object
      required: [url, interval_seconds]
      properties:
        url: { type: string }
        tiers:
          type: array
          items:
            type: string
            enum: [network, browser, vitals, lighthouse, all]
        runs: { type: integer, default: 1, maximum: 10 }
        interval_seconds:
          type: integer
          minimum: 60
          description: How often to run the test. The first run fires on the next scheduler tick (~15s).
        budget:
          $ref: '#/components/schemas/Budget'
        webhook_url: { type: string }
        profile:
          $ref: '#/components/schemas/ThrottlingProfile'

    CreateJobRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          example: https://example.com
        tiers:
          type: array
          items:
            type: string
            enum: [network, browser, vitals, lighthouse, all]
        runs:
          type: integer
          default: 1
          maximum: 10
        timeout_s:
          type: integer
          description: Per-run timeout in seconds. 0 uses the server default (GOST_TIMEOUT_S).
          minimum: 0
          maximum: 600
        webhook_url:
          type: string
        budget:
          $ref: '#/components/schemas/Budget'
        profile:
          $ref: '#/components/schemas/ThrottlingProfile'

    Result:
      type: object
      properties:
        run_index:
          type: integer
        network:
          type: object
        browser:
          type: object
        vitals:
          $ref: '#/components/schemas/VitalsResult'
        collected_at:
          type: string
          format: date-time

    VitalsResult:
      type: object
      description: >
        Lab web vitals. INP is intentionally absent — it requires real user
        interactions and cannot be measured in a lab load; TBT is the standard
        lab proxy.
      properties:
        lcp_ms: { type: number, description: Largest Contentful Paint }
        fcp_ms: { type: number, description: First Contentful Paint }
        cls:
          type: number
          description: Cumulative Layout Shift score (unitless, session-windowed).
        tbt_ms:
          type: number
          description: >
            Total Blocking Time — long-task time beyond 50ms per task after
            FCP, over the lab observation window (navigation to load + 3s).

    RUMEventRequest:
      type: object
      required: [url, name, value]
      properties:
        url: { type: string, description: Absolute http(s) page URL the metric was measured on. }
        name:
          type: string
          enum: [LCP, CLS, INP, FCP, TTFB]
        value:
          type: number
          minimum: 0
          exclusiveMaximum: 600000

    RUMSummary:
      type: object
      properties:
        url: { type: string }
        window_hours: { type: integer }
        metrics:
          type: object
          description: Keyed by metric name (LCP, CLS, INP, FCP, TTFB).
          additionalProperties:
            $ref: '#/components/schemas/MetricStats'

paths:
  /v1/health:
    get:
      summary: Liveness check
      responses:
        '200':
          description: Server is alive
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /v1/ready:
    get:
      summary: Readiness check
      responses:
        '200':
          description: Server is ready (plain-text body "READY")
          content:
            text/plain:
              schema:
                type: string
                example: READY
        '503':
          description: Database not ready

  /v1/jobs:
    post:
      summary: Submit a new test job
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  status: { type: string }
        '400':
          description: Invalid request (bad URL, invalid webhook_url, invalid budget, or runs > 10)
        '401':
          description: Unauthorized

    get:
      summary: List recent jobs
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'

  /v1/jobs/{id}:
    get:
      summary: Get job status and results
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job details and result payload
        '404':
          description: Job not found

    delete:
      summary: Delete a job (cancels it first if still pending)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Job deleted (its results and webhook deliveries are removed too)

  /v1/schedules:
    post:
      summary: Create a recurring monitor
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '201':
          description: Schedule created (enabled, first run on the next tick)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Invalid request (bad URL/tier/budget, interval_seconds < 60, or runs > 10)
        '401':
          description: Unauthorized

    get:
      summary: List schedules
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: All schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'

  /v1/schedules/{id}:
    get:
      summary: Get a schedule
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          description: Schedule not found

    patch:
      summary: Enable or disable a schedule
      description: The only mutable field is `enabled`; delete and recreate for anything else.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        '200':
          description: Updated schedule
        '400':
          description: Body was not {"enabled" true|false}
        '404':
          description: Schedule not found

    delete:
      summary: Delete a schedule (historical jobs are kept)
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Schedule deleted
        '404':
          description: Schedule not found

  /metrics:
    get:
      summary: Prometheus metrics (text exposition format)
      description: >
        Requires the API key — configure your Prometheus scrape job to send the
        X-API-Key header. Exposes gost_jobs_total, gost_job_duration_seconds,
        gost_queue_depth, gost_webhook_deliveries_total, gost_scheduler_runs_total,
        gost_retention_purged_total, and gost_last_metric_ms (scheduled URLs only).
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Metrics in Prometheus text format
          content:
            text/plain: {}
        '401':
          description: Unauthorized

  /v1/history:
    get:
      summary: Get performance aggregations for a URL
      security:
        - ApiKeyAuth: []
      parameters:
        - name: url
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Historical statistics (avg/p50/p75/p95 per metric)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/History'

  /v1/rum:
    post:
      summary: Ingest a real-user metric beacon
      description: >
        Public (no API key) so browsers can beacon directly, but disabled with
        404 unless GOST_RUM_ORIGINS is configured. CORS-checked against that
        allow-list, body capped at 8 KB, values validated, and globally
        rate-limited. Send with navigator.sendBeacon using the web-vitals
        library (see README "Real User Monitoring").
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RUMEventRequest'
      responses:
        '204':
          description: Event stored
        '400':
          description: Invalid metric name, value, or URL
        '403':
          description: Origin not in the allow-list
        '404':
          description: Endpoint disabled (GOST_RUM_ORIGINS not configured)
        '429':
          description: Rate limit exceeded

  /v1/rum/summary:
    get:
      summary: Field-data percentiles for a URL
      security:
        - ApiKeyAuth: []
      parameters:
        - name: url
          in: query
          required: true
          schema: { type: string }
        - name: window_h
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 2160, default: 24 }
          description: Look-back window in hours.
      responses:
        '200':
          description: Per-metric count/avg/p50/p75/p95 from real-user beacons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RUMSummary'
        '400':
          description: Missing url or invalid window_h
        '401':
          description: Unauthorized